IDEA-65467 check classes.dex and *.res.apk file exists during generation instead of preparation

+++ plugins/android/src/org/jetbrains/android/compiler/tools/AndroidApkBuilder.java
@@ -184,8 +184,20 @@ public class AndroidApkBuilder {
         return result;
       }
 
+      if (!new File(apkPath).exists()) {
+        result.get(CompilerMessageCategory.ERROR).add("File " + apkPath + " not found. Try to rebuild project");
+        return result;
+      }
+
+      File dexEntryFile = new File(dexPath);
+      if (!dexEntryFile.exists()) {
+        result.get(CompilerMessageCategory.ERROR).add("File " + dexEntryFile.getPath() + " not found. Try to rebuild project");
+        return result;
+      }
+
       fos = new FileOutputStream(outputApk);
       SignedJarBuilder builder = new SignedJarBuilder(fos, key, certificate);
+
       FileInputStream fis = new FileInputStream(apkPath);
       try {
         builder.writeZip(fis, null);
@@ -193,8 +205,8 @@ public class AndroidApkBuilder {
       finally {
         fis.close();
       }
-      File entryFile = new File(dexPath);
-      builder.writeFile(entryFile, AndroidUtils.CLASSES_FILE_NAME);
+
+      builder.writeFile(dexEntryFile, AndroidUtils.CLASSES_FILE_NAME);
       for (VirtualFile sourceRoot : sourceRoots) {
         writeStandardSourceFolderResources(builder, sourceRoot, sourceRoot, new HashSet<VirtualFile>(), new HashSet<String>());
       }
